WIP dump - secondary indexes
Testing raw mnesia:dirty_select operations shows that mnesia does *not* use secondary indexes unless the value is set in the head of the match.
I found this odd so decided to see what I could do to use the indexes in all select operations based on the full match spec.

This initial work create/deletes/updates the index reference tables,
will try and use an index table for a select before hitting the primary
Use range-based selects where it can based on the guards